Add RepoUtil / Implement User Service interface

+++ src/main/java/org/waterwood/waterfunservice/repository/PermissionRepo.java
@@ -3,5 +3,5 @@ package org.waterwood.waterfunservice.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.waterwood.waterfunservice.entity.permission.Permission;
 
-public interface PermissionRepo extends JpaRepository<Permission, Long> {
+public interface PermissionRepo extends JpaRepository<Permission, Integer> {
 }
